fix: move epoch_service tests into their own file
